removed user_manager_client, renamed jenkins_client package
