--- user_manager_client/response.go
@@ -1,48 +1,0 @@
-package user_manager_client
-
-import (
-	"encoding/json"
-	"io"
-	"net/http"
-)
-
-// Response is an API response.
-// This wraps the standard http.Response.
-type Response struct {
-	*http.Response
-}
-
-// newResponse creates a new Response for the provided http.Response.
-func newResponse(r *http.Response) *Response {
-	response := &Response{Response: r}
-	return response
-}
-
-// Do sends an API request and returns the API response.  The API response is
-// JSON decoded and stored in the value pointed to by v, or returned as an
-// error if an API error has occurred.  If v implements the io.Writer
-// interface, the raw response body will be written to v, without attempting to
-// first decode it.
-func (c *Client) Do(req *http.Request, v interface{}) (*Response, error) {
-	resp, err := c.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	defer resp.Body.Close()
-
-	response := newResponse(resp)
-
-	if v != nil {
-		if w, ok := v.(io.Writer); ok {
-			io.Copy(w, resp.Body)
-		} else {
-			err = json.NewDecoder(resp.Body).Decode(v)
-			if err == io.EOF {
-				err = nil // ignore EOF errors caused by empty response body
-			}
-		}
-	}
-
-	return response, err
-}
